:recycle: [REFACTOR] 테스트코드 수정
구매 번호 중복 확인, 기존 테스트코드 DisplayName 다듬기
